pyproject.toml
[project]
name = "stockpilot-investment-app"
version = "0.1.0"
description = "Full data engineering pipeline for ingesting, processing, and analyzing stock market data."
authors = [
    {name = "Ken Nagata",email = "56562713+ken-nagata@users.noreply.github.com"}
]
package-mode = false

requires-python = ">=3.12,<3.13 || >3.13,<3.14" #">=3.12,<3.14"
dependencies = [
    "yfinance (>=0.2.66,<0.3.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "google-cloud-storage (>=3.6.0,<4.0.0)",
    "google-cloud-bigquery (>=3.38.0,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "streamlit (>=1.51.0,<2.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "plotly (>=6.4.0,<7.0.0)",
    "ta (>=0.11.0,<0.12.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "gcsfs (>=2025.10.0,<2026.0.0)",
    "apache-airflow (>=3.1.3,<4.0.0)",
    "apache-airflow-providers-google (>=19.0.0,<20.0.0)",
    "apache-airflow-providers-fab (>=3.0.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
